# public/swagger.json
{"swagger":"2.0","info":{"title":"SenseHome-Technologies API Users","version":"1.0.0","description":"The purpose of the API developed is to manage the users of a system, including registration, authentication and information management functions."},"paths":{"/api/user/delete":{"delete":{"tags":["User"],"summary":"This endpoint allows a user to be deleted from the database using a JWT token and user ID.","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Id of the user to be deleted"}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/DeleteUser"}}}}}},"/api/user/edit":{"put":{"tags":["User"],"summary":"This endpoint allows a user to edit their information in the database.","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"This endpoint allows a user to edit their information in the database.","required":true,"schema":{"$ref":"#/definitions/Edit_User"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/Edit_User"}}}}}},"/api/user/forgot-password":{"post":{"tags":["User"],"summary":"This endpoint allows a user to request a password reset link to be sent to their email.","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"This endpoint allows a user to request a password reset link to be sent to their email.","required":true,"schema":{"$ref":"#/definitions/Forgot_Password"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/Forgot_Password"}}}}}},"/api/user/register":{"post":{"tags":["User"],"summary":"This endpoint allows a user to register by providing their username, email, and password. If the credentials are valid, a JWT token is returned to the user.","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"This endpoint allows a user to register by providing their username, email, and password. If the credentials are valid, a JWT token is returned to the user.","required":true,"schema":{"$ref":"#/definitions/Register_User"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/Register_User"}}}}}},"/api/user/reset-password":{"post":{"tags":["User"],"summary":"This endpoint allows a user to reset their password by providing a valid JWT token and a new password.","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"This endpoint allows a user to reset their password by providing a valid JWT token and a new password.","required":true,"schema":{"$ref":"#/definitions/ResetPassword"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/ResetPassword"}}}}}},"/api/user/login":{"post":{"tags":["User"],"summary":"This endpoint allows a user to log in by providing their email and password. If the credentials are valid, a JWT token is returned to the user.","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"This endpoint allows a user to log in by providing their email and password. If the credentials are valid, a JWT token is returned to the user.","required":true,"schema":{"$ref":"#/definitions/UserLogin"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/UserLogin"}}}}}},"/api/user/verifyCode":{"post":{"tags":["User"],"summary":"This endpoint verifies the verification code provided by the user.","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"This endpoint verifies the verification code provided by the user.","required":true,"schema":{"$ref":"#/definitions/Verify_Code"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/Verify_Code"}}}}}}},"definitions":{"DeleteUser":{"properties":{"id":{"type":"string","description":"Id of the user to be deleted"},"message":{"type":"string","description":"User deleted successfully"}},"required":["id","message"]},"Edit_User":{"properties":{"id":{"type":"string","description":"Id of the user"},"username":{"type":"string","description":"New username of the user"},"email":{"type":"string","description":"New email of the user"},"profilephoto":{"type":"string","description":"New profile photo of the user"},"phonetoken":{"type":"string","description":"New phone token of the user"},"message":{"type":"string","description":"User edited successfully"}},"required":["id","username","email","profilephoto","phonetoken","message"]},"Forgot_Password":{"properties":{"email":{"type":"string","description":"Email of the user"},"message":{"type":"string","description":"Password reset link sent to the user via email"}},"required":["email","message"]},"Register_User":{"properties":{"username":{"type":"string","description":"Username of the user"},"password":{"type":"string","description":"Password of the user"},"email":{"type":"string","description":"Email of the user"},"message":{"type":"string","description":"User registered successfully"}},"required":["username","password","email","message"]},"ResetPassword":{"properties":{"password":{"type":"string","description":"New password of the user"},"message":{"type":"string","description":"Password updated successfully"}},"required":["password","message"]},"UserLogin":{"properties":{"email":{"type":"string","description":"Email of the user"},"password":{"type":"string","description":"Password of the user"},"message":{"type":"string","description":"User logged in successfully"},"token":{"type":"string","description":"JWT token for the user"}},"required":["email","password","message","token"]},"Verify_Code":{"properties":{"email":{"type":"string","description":"Email of the user"},"verificationcode":{"type":"string","description":"Verification code provided by the user"},"message":{"type":"string","description":"Verification code is valid"},"token":{"type":"string","description":"JWT token for the user"}},"required":["email","verificationcode","message","token"]}}}